// src/types/matrixEnjinV603.ts
export type Perbill = number
// ...
export interface Listing {
    seller: AccountId32
    makeAssetId: AssetId
    takeAssetId: AssetId
    amount: bigint
    price: bigint
    minTakeValue: bigint
    feeSide: FeeSide
    creationBlock: number
    deposit: bigint
    salt: Bytes
    data: ListingData
    state: ListingState
}

export type ListingState = ListingState_Auction | ListingState_FixedPrice

export interface ListingState_Auction {
    __kind: 'Auction'
    value: AuctionState
}

export interface ListingState_FixedPrice {
    __kind: 'FixedPrice'
    amountFilled: bigint
}

export interface AuctionState {
    highBid?: Bid | undefined
}

export interface Bid {
    bidder: AccountId32
    price: bigint
}

export type ListingData = ListingData_Auction | ListingData_FixedPrice

export interface ListingData_Auction {
    __kind: 'Auction'
    value: AuctionData
}

export interface ListingData_FixedPrice {
    __kind: 'FixedPrice'
}

export interface AuctionData {
    startBlock: number
    endBlock: number
}

export type FeeSide = FeeSide_Make | FeeSide_NoFee | FeeSide_Take

export interface FeeSide_Make {
    __kind: 'Make'
}

export interface FeeSide_NoFee {
    __kind: 'NoFee'
}

export interface FeeSide_Take {
    __kind: 'Take'
}
// ...
export interface AssetId {
    collectionId: bigint
    tokenId: bigint
}
// ...
export type H256 = Bytes

export interface PerDispatchClass {
    normal: Weight
    operational: Weight
    mandatory: Weight
}

export const PerDispatchClass: sts.Type<PerDispatchClass> = sts.struct(() => {
    return {
        normal: Weight,
        operational: Weight,
        mandatory: Weight,
    }
})

export type AccountId32 = Bytes
// ...
export const AuctionData: sts.Type<AuctionData> = sts.struct(() => {
    return {
        startBlock: sts.number(),
        endBlock: sts.number(),
    }
})
// ...
export const AssetId: sts.Type<AssetId> = sts.struct(() => {
    return {
        collectionId: sts.bigint(),
        tokenId: sts.bigint(),
    }
})
// ...
export const Bid: sts.Type<Bid> = sts.struct(() => {
    return {
        bidder: AccountId32,
        price: sts.bigint(),
    }
})

export const Listing: sts.Type<Listing> = sts.struct(() => {
    return {
        seller: AccountId32,
        makeAssetId: AssetId,
        takeAssetId: AssetId,
        amount: sts.bigint(),
        price: sts.bigint(),
        minTakeValue: sts.bigint(),
        feeSide: FeeSide,
        creationBlock: sts.number(),
        deposit: sts.bigint(),
        salt: sts.bytes(),
        data: ListingData,
        state: ListingState,
    }
})

export const ListingState: sts.Type<ListingState> = sts.closedEnum(() => {
    return {
        Auction: AuctionState,
        FixedPrice: sts.enumStruct({
            amountFilled: sts.bigint(),
        }),
    }
})

export const AuctionState: sts.Type<AuctionState> = sts.struct(() => {
    return {
        highBid: sts.option(() => Bid),
    }
})

export const ListingData: sts.Type<ListingData> = sts.closedEnum(() => {
    return {
        Auction: AuctionData,
        FixedPrice: sts.unit(),
    }
})

export const FeeSide: sts.Type<FeeSide> = sts.closedEnum(() => {
    return {
        Make: sts.unit(),
        NoFee: sts.unit(),
        Take: sts.unit(),
    }
})
// ...
export const Perbill = sts.number()
// ...
export const H256 = sts.bytes()

export const AccountId32 = sts.bytes()